Script draft and scene manager
Create files for the core of the game
created a first draft of the first arc of the game

Worked on scene manager. controls scenes, just wanted to have something to start with

Going to start on the assets, when start scene.
will move onto the the json files and their code for inspections and dialogue