System manger and Start scene done
finished the system manger that handles loading scenes.

Start scene is done. So it loads two buttons and takes you to the next scene

created shells for bedroom, the first level.

Updated game.cpp so it loads scene manger which loads start scene

cometed the character shell. Going to move it in bedroom, but eventually will make global stats. that is the plan, from what I make.

updated assets.

next goal is:
- finish the next scene
- start inpection script