script arc 1 draft story
This is the most complete Arc 1.

The game will be spilt into three arcs

arc1 is loose. its a tutorial and introduction to the systems without it being overbearing.

Going to start arc 2, but I wanted to commit.

The goal is to finish Arc 2, then complete the system and start the dialogue system. tonight.

I was thinking of leaving arc 3 a bit last. Mostly because I don't want to make it too long or short